requirements.txt
backports.ssl-match-hostname==3.4.0.2
docutils==0.12
lockfile==0.10.2
python-daemon==2.0.5
PyYAML==3.11
requests==2.7.0
six==1.9.0
slackclient==0.15
websocket-client==0.32.0
wheel==0.24.0
